Refactor ImageOverlay to ImageLayer for more consistent usage of 'overlay'
